feat: sync plaid (#59)
* clean: rename func

* fix: jinja misuse

* db: migrate transactions

* pending status
* item_id (where the transaction is coming from)

* clean: style

* conf: add env to example

* ui: 500 error page

* ui: minor

* api: batch create

* fix: nullable category

* feat: mark log in status

* db: migrate plaid item

* refactor: dataclass for result structure

* deps: limit deps from breaking changes

* feat: background service for transaction pulling

* feat: add background tasks

* fix: no-no

* refactor: organize plaid specifics

* git: ignore scripts

* db: manage plaid accounts

* db: manage plaid transactions

* feat: sync with plaid

* conf: add rich for development

* fix: wrong id

* ui: visual uncategorized

* fix: coalesce plaid sync task

* ui: store user preference for theme

* fix: limit what is sent to context

* refactor: clear module name

* fix: validation error

* feat: timeout session

* clean: remove prints/comments

* test: delete dummy tests

* feat: store request info

* fix: missing field

* refactor: rename var

* fix: only retrieve active items

* feat: sync transactions after succesful link

* feat: plaid remove item

* clean: remove unused env var

* fix: bad endpoint

* feat: automatic category for plaid

* db: add user_id to plaid transactions

* fix: bug selecting page

* fix: transaction ui pagination

* test: fix

* fix: keep transaction page state

* db: add default categories

* git: ignore all almond databases

+++ src/almonds/api/plaid.py
@@ -1,21 +1,25 @@
+from threading import Thread
 from uuid import UUID
 
 from flask import Blueprint, request, session
 
-from almonds.crud import plaid_item as crud_plaid_item
+import almonds.crud.user as crud_user
+import almonds.services.plaid.core as plaid_core
+from almonds.crud.plaid import plaid_item as crud_plaid_item
 from almonds.crypto.cryptograph import Cryptograph
-from almonds.schemas.plaid_item import PlaidItemBase
-from almonds.services.plaid import core
+from almonds.schemas.plaid.plaid_item import PlaidItemBase
+from almonds.services import plaid_sync
 
 plaid_bp = Blueprint("plaid", __name__)
 
 
 @plaid_bp.route("/createLinkToken")
 def create_link_token():
-    user_id = (session.get("user_id", None),)
+    user_id = session.get("user_id", None)
     if user_id is None:
         return {"error": "user_id not set"}
-    link_token = core.create_link_token(user_id)
+
+    link_token = plaid_core.create_link_token(user_id)
     return {"link_token": link_token}
 
 
@@ -25,7 +29,7 @@ def exchange_public_token():
         return {"public_token_exchange": "ERROR"}
 
     body = request.get_json()
-    resp = core.exchange_public_token(body["public_token"])
+    resp = plaid_core.exchange_public_token(session["user_id"], body["public_token"])
 
     # Encrypt the data.
     crypto = Cryptograph()
@@ -35,8 +39,15 @@ def exchange_public_token():
         access_token=crypto.encrypt(resp["access_token"]),
         item_id=crypto.encrypt(resp["item_id"]),
     )
-    print(f"Plaid Item: {item}")
     crud_plaid_item.create_item(item)
+
+    # start a sync in the background
+    Thread(
+        target=plaid_sync.sync_user,
+        args=(crud_user.get_user_by_id(session["user_id"]),),
+        kwargs={"cryptograph": crypto},
+        daemon=True,
+    ).start()
 
     return {"public_token_exchange": "complete"}
 
@@ -50,6 +61,16 @@ def is_account_connected():
 def delete_access_token():
     body = request.get_json()
     item_id = UUID(body["id"])
+    item = crud_plaid_item.get_item(item_id)
+
+    # validate
+    if item is None:
+        return {"status": "not found"}, 404
+
     crud_plaid_item.delete_item(item_id)
+
+    crypto = Cryptograph()
+
+    plaid_core.remove_item(item.user_id, crypto.decrypt(item.access_token))
 
     return {"status": "deleted"}
